chore(checkoutv2): remove Checkout page components
