+++ apps/store/src/app/[locale]/new/checkout/CheckoutForm.tsx
@@ -4,8 +4,8 @@ import { useApolloClient } from '@apollo/client'
 import { useTranslation } from 'next-i18next'
 import { useState } from 'react'
 import { Text, Button, BankIdIcon, yStack } from 'ui'
+import { useHandleSubmitCheckout } from '@/app/[locale]/new/checkout/useHandleSubmitCheckout'
 import { useAsyncRouterPush } from '@/appComponents/useAsyncRouterPush'
-import { useHandleSubmitCheckout } from '@/components/CheckoutPage/useHandleSubmitCheckout'
 import * as FullscreenDialog from '@/components/FullscreenDialog/FullscreenDialog'
 import { PersonalNumberField } from '@/components/PersonalNumberField/PersonalNumberField'
 import { TextField } from '@/components/TextField/TextField'
