--- apps/store/src/components/CheckoutPage/CheckoutPage.constants.ts
@@ -1,9 +1,0 @@
-export enum FormElement {
-  Email = 'email',
-  FirstName = 'firstName',
-  LastName = 'lastName',
-}
-
-export enum QueryParam {
-  ExpandCart = 'expand',
-}
