--- apps/store/src/components/CheckoutPage/PageDebugDialog.tsx
@@ -1,5 +1,0 @@
-'use client'
-
-import { DefaultDebugDialog } from '@/components/DebugDialog/DefaultDebugDialog'
-
-export const PageDebugDialog = DefaultDebugDialog
