+++ apps/store/src/features/widget/SignPage.tsx
@@ -8,8 +8,8 @@ import { useRouter } from 'next/router'
 import { useTranslation } from 'next-i18next'
 import { type FormEvent, type MouseEvent, type PropsWithChildren, useState } from 'react'
 import { BankIdIcon, Button, CheckIcon, Heading, mq, Space, Text, theme } from 'ui'
-import { FormElement } from '@/components/CheckoutPage/CheckoutPage.constants'
-import { useHandleSubmitCheckout } from '@/components/CheckoutPage/useHandleSubmitCheckout'
+import { FormElement } from '@/app/[locale]/new/checkout/CheckoutPage.constants'
+import { useHandleSubmitCheckout } from '@/app/[locale]/new/checkout/useHandleSubmitCheckout'
 import * as FullscreenDialog from '@/components/FullscreenDialog/FullscreenDialog'
 import * as GridLayout from '@/components/GridLayout/GridLayout'
 import { PersonalNumberField } from '@/components/PersonalNumberField/PersonalNumberField'
